- refactor DB model - clean code
